Support for associated functions and methods
