+++ src/peripheral/plic.rs
@@ -2,11 +2,8 @@
 //!
 //! Specification: https://github.com/riscv/riscv-plic-spec/blob/master/riscv-plic.adoc
 
-use super::PLIC;
-use crate::{
-    interrupt::{InterruptNumber, PriorityNumber},
-    register::mie,
-};
+pub use super::PLIC;
+use crate::register::mie;
 use core::ops::Deref;
 use volatile_register::{RO, RW};
 
@@ -51,12 +48,8 @@ pub struct ContextState {
 impl<const BASE: usize, const CONTEXT: usize> PLIC<BASE, CONTEXT> {
     /// Sets the Machine External Interrupt bit of the [`crate::register::mie`] CSR.
     /// This bit must be set for the PLIC to trigger machine external interrupts.
-    ///
-    /// # Safety
-    ///
-    /// Enabling machine external interrupts globally can break mask-based critical sections.
     #[inline]
-    pub unsafe fn mext_enable() {
+    pub fn enable() {
         // SAFETY: atomic CSRRS instruction with no side effects
         unsafe { mie::set_mext() };
     }
@@ -64,19 +57,24 @@ impl<const BASE: usize, const CONTEXT: usize> PLIC<BASE, CONTEXT> {
     /// Clears the Machine External Interrupt bit of the [`crate::register::mie`] CSR.
     /// When cleared, the PLIC does not trigger machine external interrupts.
     #[inline]
-    pub fn mext_disable() {
+    pub fn disable() {
         // SAFETY: atomic CSRRC instruction with no side effects
         unsafe { mie::clear_mext() };
     }
 
     /// Returns the priority level associated to a given interrupt source.
-    /// If priority level is 0 (i.e., "never interrupt"), it returns `None`.
     #[inline]
-    pub fn get_priority<I: InterruptNumber, P: PriorityNumber>(source: I) -> P {
+    pub fn priority<I: InterruptNumber, P: PriorityNumber>(source: I) -> P {
         let source = usize::from(source.number());
         // SAFETY: atomic read with no side effects
         let priority = unsafe { (*Self::PTR).priority[source].read() } as _;
         P::try_from(priority).unwrap()
+    }
+
+    /// Getter method for the priority level associated to a given interrupt source.
+    #[inline(always)]
+    pub fn get_priority<I: InterruptNumber, P: PriorityNumber>(&self, source: I) -> P {
+        Self::priority(source)
     }
 
     /// Sets the priority level of a given interrupt source.
@@ -90,7 +88,11 @@ impl<const BASE: usize, const CONTEXT: usize> PLIC<BASE, CONTEXT> {
     ///
     /// Changing priority levels can break priority-based critical sections and compromise memory safety.
     #[inline]
-    pub unsafe fn set_priority<I: InterruptNumber, P: PriorityNumber>(source: I, priority: P) {
+    pub unsafe fn set_priority<I: InterruptNumber, P: PriorityNumber>(
+        &mut self,
+        source: I,
+        priority: P,
+    ) {
         let source = usize::from(source.number());
         let priority = priority.number().into();
         // SAFETY: atomic write with no side effects
@@ -119,17 +121,12 @@ impl<const BASE: usize, const CONTEXT: usize> PLIC<BASE, CONTEXT> {
 
     /// Enables an interrupt source for the PLIC context.
     ///
-    /// # Note
-    ///
-    /// This method performs a read-modify-write operation.
-    /// That is why it is a method instead of an associated function.
-    ///
     /// # Safety
     ///
-    /// Non-atomic operations may lead to undefined behavior.
-    /// Enabling an interrupt source can break mask-based critical sections.
+    /// It performs non-atomic read-modify-write operations, which may lead to undefined behavior.
+    /// Additionally, Enabling an interrupt source can break mask-based critical sections.
     #[inline]
-    pub unsafe fn interrupt_enable<I: InterruptNumber>(&mut self, source: I) {
+    pub unsafe fn enable_interrupt<I: InterruptNumber>(&mut self, source: I) {
         let source = usize::from(source.number());
         let mask: u32 = 1 << (source % MAX_FLAGS_WORDS);
         self.enables[CONTEXT][source / MAX_FLAGS_WORDS].modify(|value| value | mask);
@@ -137,27 +134,28 @@ impl<const BASE: usize, const CONTEXT: usize> PLIC<BASE, CONTEXT> {
 
     /// Disables an interrupt source for the PLIC context.
     ///
-    /// # Note
-    ///
-    /// This method performs a read-modify-write operation.
-    /// That is why it is a method instead of an associated function.
-    ///
     /// # Safety
     ///
-    /// Non-atomic operations may lead to undefined behavior.
+    /// It performs non-atomic read-modify-write operations, which may lead to undefined behavior.
     #[inline]
-    pub unsafe fn interrupt_disable<I: InterruptNumber>(&mut self, source: I) {
+    pub unsafe fn disable_interrupt<I: InterruptNumber>(&mut self, source: I) {
         let source = usize::from(source.number());
         let mask: u32 = 1 << (source % MAX_FLAGS_WORDS);
         self.enables[CONTEXT][source / MAX_FLAGS_WORDS].modify(|value| value & !mask);
     }
 
-    /// Gets the priority threshold for for the PLIC context.
+    /// Returns the priority threshold of the PLIC context.
     #[inline]
-    pub fn get_threshold<P: PriorityNumber>() -> P {
+    pub fn threshold<P: PriorityNumber>() -> P {
         // SAFETY: atomic read with no side effects
         let priority = unsafe { (*Self::PTR).states[CONTEXT].threshold.read() } as _;
         P::try_from(priority).unwrap()
+    }
+
+    /// Getter method for the priority threshold of the PLIC context.
+    #[inline(always)]
+    pub fn get_threshold<P: PriorityNumber>(&self) -> P {
+        Self::threshold()
     }
 
     /// Sets the priority threshold for for the PLIC context.
@@ -166,13 +164,14 @@ impl<const BASE: usize, const CONTEXT: usize> PLIC<BASE, CONTEXT> {
     ///
     /// Unmasking an interrupt source can break mask-based critical sections.
     #[inline]
-    pub unsafe fn set_threshold<P: PriorityNumber>(priority: P) {
+    pub unsafe fn set_threshold<P: PriorityNumber>(&mut self, priority: P) {
         let priority = priority.number().into();
         // SAFETY: atomic write with no side effects
         (*Self::PTR).states[CONTEXT].threshold.write(priority);
     }
 
     /// Claims the number of a pending interrupt for for the PLIC context.
+    /// If no interrupt is pending for this context, it returns [`None`].
     #[inline]
     pub fn claim<I: InterruptNumber>() -> Option<I> {
         // SAFETY: atomic read with no side effects
@@ -184,6 +183,11 @@ impl<const BASE: usize, const CONTEXT: usize> PLIC<BASE, CONTEXT> {
     }
 
     /// Marks a pending interrupt as complete from for the PLIC context.
+    ///
+    /// # Note
+    ///
+    /// If the source ID does not match an interrupt source that is
+    /// currently enabled for the target, the completion is silently ignored.
     #[inline]
     pub fn complete<I: InterruptNumber>(source: I) {
         let source = source.number().into();
@@ -206,14 +210,14 @@ impl<const BASE: usize, const CONTEXT: usize> PLIC<BASE, CONTEXT> {
     ///
     /// # Safety
     ///
-    /// Non-atomic operations may lead to undefined behavior.
-    #[inline]
+    /// It performs non-atomic read-modify-write operations, which may lead to undefined behavior.
+    #[inline(always)]
     pub unsafe fn reset<I: InterruptNumber, P: PriorityNumber>(&mut self) {
-        Self::set_threshold(P::try_from(P::MAX_PRIORITY_NUMBER).unwrap());
+        self.set_threshold(P::try_from(P::MAX_PRIORITY_NUMBER).unwrap());
         let no_interrupt = P::try_from(0).unwrap();
         for source in (1..=I::MAX_INTERRUPT_NUMBER).filter_map(|n| I::try_from(n).ok()) {
-            self.interrupt_disable(source);
-            Self::set_priority(source, no_interrupt);
+            self.disable_interrupt(source);
+            self.set_priority(source, no_interrupt);
         }
     }
 }
@@ -228,3 +232,66 @@ impl<const BASE: usize, const CONTEXT: usize> Deref for PLIC<BASE, CONTEXT> {
 }
 
 unsafe impl<const BASE: usize, const CONTEXT: usize> Send for PLIC<BASE, CONTEXT> {}
+
+/// Trait for enums of interrupt numbers.
+///
+/// This trait should be implemented by a peripheral access crate (PAC)
+/// on its enum of available external interrupts for a specific device.
+/// Each variant must convert to a `u16` of its interrupt number.
+///
+/// # Note
+///
+/// Recall that the interrupt number `0` is reserved as "no interrupt".
+///
+/// # Safety
+///
+/// This trait must only be implemented on enums of external interrupts. Each
+/// enum variant must represent a distinct value (no duplicates are permitted),
+/// and must always return the same value (do not change at runtime).
+/// All the interrupt numbers must be less than or equal to `MAX_INTERRUPT_NUMBER`.
+/// `MAX_INTERRUPT_NUMBER` must coincide with the highest allowed interrupt number.
+///
+/// These requirements ensure safe nesting of critical sections.
+pub unsafe trait InterruptNumber: Copy {
+    /// Highest number assigned to an interrupt source.
+    const MAX_INTERRUPT_NUMBER: u16;
+
+    /// Converts an interrupt source to its corresponding number.
+    fn number(self) -> u16;
+
+    /// Tries to convert a number to a valid interrupt source.
+    /// If the conversion fails, it returns an error with the number back.
+    fn try_from(value: u16) -> Result<Self, u16>;
+}
+
+/// Trait for enums of interrupt priority numbers.
+///
+/// This trait should be implemented by a peripheral access crate (PAC)
+/// on its enum of available priority numbers for a specific device.
+/// Each variant must convert to a `u8` of its priority level.
+///
+/// # Note
+///
+/// Recall that the priority number `0` is reserved as "never interrupt".
+///
+/// # Safety
+///
+/// This trait must only be implemented on enums of priority levels. Each
+/// enum variant must represent a distinct value (no duplicates are permitted),
+/// and must always return the same value (do not change at runtime).
+/// There must be a valid priority number set to 0 (i.e., never interrupt).
+/// All the priority level numbers must be less than or equal to `MAX_PRIORITY_NUMBER`.
+/// `MAX_PRIORITY_NUMBER` must coincide with the highest allowed priority number.
+///
+/// These requirements ensure safe nesting of critical sections.
+pub unsafe trait PriorityNumber: Copy {
+    /// Number assigned to the highest priority level.
+    const MAX_PRIORITY_NUMBER: u8;
+
+    /// Converts a priority level to its corresponding number.
+    fn number(self) -> u8;
+
+    /// Tries to convert a number to a valid priority level.
+    /// If the conversion fails, it returns an error with the number back.
+    fn try_from(value: u8) -> Result<Self, u8>;
+}
